Log to memory by default
Instead of logging to file by default log to memory instead. To be able
to look at the logs add a command LspLog to print. Currently the max
amount of log records are 1000, which is probably more than enough while
still not taking that much memory. The custom logging to memory is done
to avoid the logs to end up on the vim command-line, interupting the
user (potentially also overwriting current messages, e.g. as displayed
by diagnostics).